Moved parseEdgeFile to seperate class.
